Update rtl/ltr without reloading the page
