+++ client/components/tinymce/index.jsx
@@ -152,9 +152,9 @@ mentionsPlugin();
 PLUGINS.push( 'wpcom/mentions' );
 
 const CONTENT_CSS = [
-	window.app.tinymceWpSkin,
+	window.app.urls[ 'tinymce/skins/wordpress/wp-content.css' ],
 	'//s1.wp.com/wp-includes/css/dashicons.css',
-	window.app.tinymceEditorCss,
+	window.app.urls[ 'editor.css' ],
 	'//fonts.googleapis.com/css?family=Noto+Serif:400,400i,700,700i&subset=cyrillic,cyrillic-ext,greek,greek-ext,latin-ext,vietnamese',
 ];
